Create merge_sort.java
The following code is on Merge sort which is successfully compiled and run. Merge sort is one of the best algorithms when comes to time complexity, O(nlogn). It is used in Inversion Count Problem. We can use it in External Sorting as well. Please feel free to suggest changes.

Thank You!